requirements.txt
asgiref==3.7.2
asyncpg==0.29.0
asyncpgsa==0.27.1
autopep8==2.0.4
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.6
Flask==3.0.0
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.2
psycopg2==2.9.9
pycodestyle==2.11.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.1
